refactor: implementation
